fix: explicitly use int64 everywhere required (resolve #40) chore: upgrade dependencies

+++ model/store.go
@@ -2,5 +2,5 @@ package model
 
 type StoreObject struct {
 	User   TelegramUser `json:"user"`
-	ChatId int          `json:"chat_id"`
+	ChatId int64        `json:"chat_id"`
 }
